细节优化

+++ change-api/change-api-dao/change-api-dao-order/src/main/java/me/cathub/change/api/dao/order/OrderDao.java
@@ -33,7 +33,7 @@ public interface OrderDao extends BaseDao<Order> {
      * @return
      * @throws Exception
      */
-    long countByStorehouseId(long storehouse_id, int tableIndex) throws Exception;
+    int countByStorehouseId(long storehouse_id, int tableIndex) throws Exception;
 
     /**
      * 根据借卖方获取订单列表
@@ -53,7 +53,7 @@ public interface OrderDao extends BaseDao<Order> {
      * @return
      * @throws Exception
      */
-    long countByShopkeeperId(long shopkeeper_id, int tableIndex) throws Exception;
+    int countByShopkeeperId(long shopkeeper_id, int tableIndex) throws Exception;
 
     /**
      * 根据品牌商获取订单列表
@@ -72,5 +72,5 @@ public interface OrderDao extends BaseDao<Order> {
      * @param tableIndex
      * @return
      */
-    long countBrandQuotientId(long brandQuotient_id, int tableIndex);
+    int countBrandQuotientId(long brandQuotient_id, int tableIndex);
 }
